refactor: 엔티티 ID 타입 변경

+++ src/main/java/com/findex/demo/indexData/index/domain/entity/IndexData.java
@@ -9,30 +9,29 @@ import java.time.Instant;
 
 public class IndexData {
 
-  @Id
-  @GeneratedValue
-  private Long id;
+    @Id
+    @GeneratedValue
+    private Integer id;
 
-  //@ManyToOne
-  //@JoinColumn(name = "index_info_id")
-  //private IndexInfo indexInfo;
+    //@ManyToOne
+    //@JoinColumn(name = "index_info_id")
+    //private IndexInfo indexInfo;
 
-  private Instant date;
+    private Instant date;
 
-  @Enumerated(EnumType.STRING)
-  private SourceType sourceType;
+    @Enumerated(EnumType.STRING)
+    private SourceType sourceType;
 
-  private BigDecimal openPrice;
-  private BigDecimal closePrice;
-  private BigDecimal highPrice;
-  private BigDecimal lowPrice;
+    private BigDecimal openPrice;
+    private BigDecimal closePrice;
+    private BigDecimal highPrice;
+    private BigDecimal lowPrice;
 
-  private BigDecimal changeAmount;
-  private BigDecimal changeRate;
+    private BigDecimal changeAmount;
+    private BigDecimal changeRate;
 
-  private Long volume;
-  private Long tradingValue;
-  private Long marketCap;
+    private Long volume;
+    private Long tradingValue;
+    private Long marketCap;
 
 }
-
